WIP - refactor queue and impl; add gcp library
