+++ buildSrc/src/main/groovy/stairway.java-conventions.gradle
@@ -11,6 +11,19 @@ repositories {
     google()
 }
 
+dependencies {
+    // For SpotBugs annotations (still using FindBugs annotations)
+    implementation group: 'com.google.code.findbugs', name: 'annotations', version: '3.0.1'
+
+    // For logging
+    implementation group: 'org.slf4j', name: 'slf4j-api', version: '1.7.30'
+
+    // For testing
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.5.2'
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.2'
+    testImplementation group: 'org.hamcrest', name: 'hamcrest', version: '2.2'
+}
+
 // Spotless configuration
 spotless {
     java {
