+++ stairway-gcp/build.gradle
@@ -9,15 +9,10 @@ dependencies {
     }
     implementation platform('com.google.cloud:libraries-bom:20.2.0')
 
-    implementation group: 'org.postgresql', name: 'postgresql', version: '42.2.18'
-    implementation group: 'org.apache.commons', name: 'commons-dbcp2', version: '2.8.0'
-    implementation group: 'org.apache.commons', name: 'commons-pool2', version: '2.9.0'
+    api(':stairway')
+
     implementation group: 'org.apache.commons', name: 'commons-lang3', version: '3.11'
     implementation group: 'org.apache.commons', name: 'commons-collections4', version: '4.4'
-
-    // Database migration
-    implementation group: 'org.liquibase', name: 'liquibase-core', version: '4.2.1'
-    implementation group: 'org.yaml', name: 'snakeyaml', version: '1.27'
 
     // JSON processing
     implementation('com.fasterxml.jackson.core:jackson-databind:2.12.3')
@@ -25,9 +20,6 @@ dependencies {
     implementation('com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.12.3')
     implementation('com.fasterxml.jackson.module:jackson-module-parameter-names:2.12.3')
     implementation group: 'org.openapitools', name: 'jackson-databind-nullable', version: '0.2.1'
-
-    // For HttpStatus:
-    implementation group: 'org.springframework', name: 'spring-web', version: '5.3.1'
 
     // For PubSub
     implementation group: 'com.google.cloud', name: 'google-cloud-pubsub'
