+++ stairway/src/main/java/bio/terra/stairway/Flight.java
@@ -7,6 +7,8 @@ import static bio.terra.stairway.FlightStatus.WAITING;
 import bio.terra.stairway.exception.DatabaseOperationException;
 import bio.terra.stairway.exception.RetryException;
 import bio.terra.stairway.exception.StairwayExecutionException;
+import bio.terra.stairway.impl.FlightDao;
+import bio.terra.stairway.impl.HookWrapper;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
