+++ stairway/src/main/java/bio/terra/stairway/impl/FlightDao.java
@@ -1,9 +1,20 @@
-package bio.terra.stairway;
+package bio.terra.stairway.impl;
 
-import static bio.terra.stairway.DbUtils.commitTransaction;
-import static bio.terra.stairway.DbUtils.startReadOnlyTransaction;
-import static bio.terra.stairway.DbUtils.startTransaction;
+import static bio.terra.stairway.impl.DbUtils.commitTransaction;
+import static bio.terra.stairway.impl.DbUtils.startReadOnlyTransaction;
+import static bio.terra.stairway.impl.DbUtils.startTransaction;
 
+import bio.terra.stairway.Direction;
+import bio.terra.stairway.ExceptionSerializer;
+import bio.terra.stairway.FlightContext;
+import bio.terra.stairway.FlightDebugInfo;
+import bio.terra.stairway.FlightFilter;
+import bio.terra.stairway.FlightInput;
+import bio.terra.stairway.FlightMap;
+import bio.terra.stairway.FlightState;
+import bio.terra.stairway.FlightStatus;
+import bio.terra.stairway.HookWrapper;
+import bio.terra.stairway.StepResult;
 import bio.terra.stairway.exception.DatabaseOperationException;
 import bio.terra.stairway.exception.DuplicateFlightIdException;
 import bio.terra.stairway.exception.FlightException;
