+++ stairway/src/main/java/bio/terra/stairway/impl/README.md
@@ -1,0 +1,5 @@
+# Implementation Package
+This and the underlying packages hold the implementation of Stairway.
+Public classes and methods in this package are not intended for client use.
+Only public classes and methods in the top-most stairway package are
+for use by clients.
