+++ stairway/src/main/java/bio/terra/stairway/queue/QueueMessageEnum.java
@@ -1,4 +1,4 @@
-package bio.terra.stairway;
+package bio.terra.stairway.queue;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
